wip: refactor(model): factor bounding box algorithm
This diff is a work in progress to factor the bounding box algorithm
out including tests useful to have confidence that all is WAI.

The reference issue is https://github.com/tn-aixpa/dt-model/issues/7.

This code is not complete and has been mainly useful for me to
conceptualise and understand how to implement a hotfix.

We should discuss a bit more what's the right bounding box approach
after we have released the current iteration of the demo.