feat(constraint): add the optional name field
We need a name for the Constraint to map a Constraint class to its name
and draw the corresponding median lines with a legend in overtourism.

Because the name was not previously exposed as a mandatory constructor
field, I chose to make it an optional, default empty, field. This means
that we don't break existing clients using the API, which is something
I kind of live to avoid doing unnecessarily.

While there, note that group seems to be a very view specific field,
so drop a note to investigate whether a different mapping based
arragement is possible where the view keeps a map between a model
entity and the view specific fields pertaining to it. This kind
of arrangement could ot could not be feasible, considering how we
are also cloning data structures, thus the comment mentions the
need to consider and investigate only.

This commit is part of https://github.com/tn-aixpa/overtourism/issues/9
and we would need a corresponding commit inside the overtourism
repository to take advantage of this functionality to plot medians
along with a specific legend describing them.