chore: use uv and enable github actions (#2)
This diff upgrades the repository to use uv and adds github actions for
running continuous integration tests.

The motivation for writing this diff is upgrading to `uv`, which
supports a lockfile, for precise dependencies.

To have confidence that adding `uv` is not going to break the build, I
am adding support for github actions as well.

While there, tweak the existing test to use `pytest`, so we can measure
coverage, and attempt to use codecov too.

Lastly, we add badges to the README.md file.