fix(pyproject.toml): pin to Python 3.11.11
This diff updates `pyproject.toml` to pin to the latest version
of the 3.11 Python branch, maintained until 2027-10.

We must pin to this version because that's the most recent
supported version by the @scc-digitalhub platform that we're
using at @fbk-most to run Python code.

While there, notice that `uv python install` is unnecessary in
the `test.yml`, since `uv sync --dev` already installs the
required Python version, so remove the redundant code.