feat: introduce the `engine/frontend` package (#4)
This diff introduces the `engine/frontend` package. The concern of this
package is helping out to build an abstract computation graph. The
intention of building such a graph is that of evaluating it using
distinct numerical backends (e.g., NumPy, TensorFlow).

This diff in particular focuses on introducing the `graph.py module of
the `frontend` package. Such a module is responsible for allowing to
construct an untyped computation graph.

Subsequent commits will introduce other components of the engine
piecemeal. The prototype implementation for the code in this pull
request has been developed out of tree at
https://github.com/bassosimone/yakof.

This code is in sync with
https://github.com/bassosimone/yakof/commit/bcdcecc4ef7c19459e3a527730232927b249ef47.