feat(graph): assign atomic integer ID to each node
This diff extends the graph representation to assign an atomic integer
ID to each node. In the tn-aixpa/overtoursim demo we are using UUID4 in
a similar fashion, however, using incremental atomic integers has more
advantages, including that it allows us to sort nodes by their order
of insertion. In turn, this sortability property allows us to flatten
the graph tree to the insertion order easily. And, by doing this,
we are able to evaluate the graph in its natural order without keeping
track of the correct evaluation order on the side.