feat(graph): assign atomic integer ID to each node (#6)
This diff extends the graph representation to assign an atomic integer
ID to each node. In the tn-aixpa/overtourism demo we are using UUID4 in
a similar fashion, however, using incremental atomic integers has more
advantages, including that it allows us to sort nodes by their order of
insertion. In turn, this sortability property allows us to flatten the
graph tree to the insertion order easily. And, by doing this, we are
able to evaluate the graph in its natural order without keeping track of
the correct evaluation order on the side. Note that we are using an
atomic integer here instead of using just an integer because Streamlit
uses multiple threads. Also, note that we are targeting Python 3.11,
while subsequent versions seem to have better support for atomic
integers. We'll update our code when we'll get there.