fix(tests): add __init__.py to each directory (#7)
I originally thought this was not needed with `pytest` because of its
auto-discover feature. However, it seems not having an `__init__.py`
prevents us from having equally named test files in different
directories. Conversely, with an `__init__.py`, each testing module
lives in a package, thus equally named files are not a problem anymore.