feat(engine): implement topological sorting (#8)
This diff implements topological sorting of the DAG. With
topological sorting we can rest assured that we are visting
nodes in the correct, linearized order. In turn, this
allows to cache the result of the computation naturally.

We need this feature to implement a numpy evaluator that
assumes linearized nodes and implements caching.

The order with which I am submitting these patches to the
dt_model differs from the order with which I wrote them
in [yakof](https://github.com/bassosimone/yakof). Specifically,
I originally wrote an AST evaluator with optional caching for
yakof. However, with the benefit of the hindsight, I think
we don't need such an evaluator, and rather we should instead
always use topological sorting and write much simpler code
to evaluate the DAG instead.

---------

Co-authored-by: Marco Pistore <marco.pistore@gmail.com>