fix(pyproject.toml): require python 3.11.x
Previously, we were forcing 3.11.11. However, asking for 3.11.x
is more flexible. In particular, if the system installs a Python
version in the 3.11.x line, with this change we don't need to
download a custom Python version using `uv`.

Additionally, this means that we'll automatically upgrade to a
newer version of the 3.11.x line, should that be released.

I assume there are not going to be breaking changes when upgrading
to patch Python versions, so we should still be fine.

While there, be more precise and specify constraints for the
dependencies we're using, rather than just listing them by name.

To properly select such dependencies constraints, I am using
the versions listed inside of the `uv.lock` file so that there
is not any actual functional change here.