fix: use python>=3.11 at runtime, python==3.11 when typechecking (#10)
This diff attempts to allow for a better local rapid development history
by allowing python >= 3.11 while enforcing python 3.11 type checking
through a dedicated `pyrightconfig.json` configuration file. In the
configuration file, I am opting for strict type checking, which seems
the right thing to do in the long run.

The `engine` part of the code base has been implicitly developed with
strict type checking (which seem to be Zed Editor's defaults). For the
rest of the code base, we take stock of the fact that it's currently
working as intended, and we'll proceed with incremental refactoring.
That is, we will fix typing when we'll need to update the code base,
following the boy scout rule, until the camp ends up definitely very
clean and correct in terms of typing.

The problem I am trying to solve here is the same discussed in
https://github.com/bassosimone/yakof/pull/65, which, in turn, is
described at
https://github.com/astral-sh/python-build-standalone/issues/129. The
main difference is that, rather than finding a workaround for plotting
with python==3.11.11, here I am opting to allow python>=3.11 and I rely
on type checking for enforcing a 3.11.x-compatible syntax (which is what
we need in the @scc-digitalhub platform).

While there, I am adding pyright to the CI (though failures are not
blocking for now) and I am adding more runners for currently-stable
Python versions, to ensure the test suite is working as intended with
also 3.12 and 3.13.

While there, explicitly specify the minimum version of dependencies in
`pyproject.toml`. We already had a `uv.lock`, so this change is
basically a no-op, except that the TOML file is now explicit wrt the
minimum supported versions.