feat(graph): support infix and prefix operators (#11)
This diff ports https://github.com/bassosimone/yakof/pull/54 to the
`dt-model` repository. As mentioned there, I figured that we could
reimplement the top-level `dt-model` functionality w/o using tensors and
tensor spaces, at least at the beginning.

By following this strategy, we should be able to replace sympy code with
yakof based code in a smaller number of pull requests, even though,
eventually, we know we will want to use tensor spaces and tensors.

Yet, to execute this plan, we need to extend the graph to support infix
and prefix operators, otherwise, we would need to change the original
model's code too radically, which is not a Good Thing™.

While there, tweak `pyrightconfig.json`: it turns out that I was using
`standard` rather than `strict` mode. While it would generally be nice
to write code in `strict` mode, the time is definitely not ripe for
doing this, and, for our current intents and purposes, the relative
middle ground of the `standard` mode seems a better default setting. I
suppose, in the future, we could potentially use `# pyright: strict` in
selected places, if we need that.