refactor: allow `import`-ing the molveno model
This diff splits the molveno example roughly in half:

1. the first half now lives into dt_model.examples.molveno
and can be imported by example scripts and tests

2. the second half is still inside examples and is still
an executable script we can execute for exploration

This change is motivated by a desire to write tests for the
molveno movel, to ensure we're not breaking it when we upgrade
from the original dt_model code to yakof based code.

I have also edited `pyproject.toml` to ensure we're not
installing examples when `pip install`-ing the `dt_model`
package, since we don't need doing this. We just need
the model to be `import`-able from tests and examples alike.

It's also noteworthy that splitting the molveno model
in half provides some initial insights on how the whole
codebase can, in the future, be split between code
that is immutable and defines a model and model instatiations.