chore: switch to ruff and start using it (#13)
The original `pyproject.toml` was already using `ruff` and, to some
extent, my decision to use `black` instead was a step back.

There's lots of traction behind `ruff`. It is fast and it can not only
reformat but also `lint` the code base.

Hence, with this commit, I am switching from `black` to `ruff` and I am
also reformatting the existing codebase using it.

In a subsequent commit, I will attempt to also start linting the
existing codebase, however, for now, this feels like a small enough but
forward enough step to warrant its own PR and review cycle.