chore(ci): typecheck each run using pyright (#15)
This diff ensures we typecheck the codebase in each CI run using the
pyright type checker (used by VSCode).

I also fixed some easy-to-fix type errors and added to the exclude list
the files with residual type errors.

We're going to take care of them in the future and the IDE still shows
squiggles for them anyway, which, if you ask me, is a nice feature,
because it ensures we won't forget about these files having typing
errors.