feat(molveno): write test using fixed ensemble (#18)
This diff adds a test testing the frozen Molveno model using a fixed
ensemble and ensuring the results are close to expectations computed by
evaluating the model itself.

By introducing this test, we will be able to evolve the `dt_model`
package with confidence that we won't be breaking things.

While there, I modified the semantics of `Model.reset` to be idempotent.
This change allows us to `reset` a model with confidence without knowing
its previous state. Resetting an already empty model is fine anyway.