feat: replace sympy with the engine package (#24)
This diff replaces the `sympy` based implementation with the one inside
the `engine` package. As part of this change, we need to sacrifice the
model variation. (We previously explored what substituting inside a
formula might have looked like in
https://github.com/fbk-most/dt-model/pull/24.) We're moving towards a
situation in which we're decoupling more and more the abstract model and
its abstract components its instantiation. With such a code
architecture, what we end up doing is basically keeping a dictionary
mapping abstract nodes to their initial values. Hence, a model variation
is as simple as cloning the dictionary and applying the required changes
to it.