fix(pyproject.toml): use hatchling for packaging (#32)
The whole packaging procedure is significantly simpler if we use
hatchling as opposed to setuptools. With out configuration, the packages
generated by setuptools were broken according to the

    twin check dist/*

command. Conversely, hatchling, which is more modern, looks great and
passes all the checks.

While there, extend the README.md file to explain how to build releases
and publish them on pypi.

As a future looking note, we will see to automatically package using
GitHub actions in the future.