chore: enforce numpy documenting syntax (#35)
I am going to consider this activity as an experiment to understand
whether enforcing numpy docstring syntax is actually good for us or too
strict. Speaking in purely practical terms, doing this right now is good
for me. I have a separate codebase (bassosimone/yakof) that I need to
keep in sync, so I am trying to reduce the code duplication. In turn,
predictable formatting helps me to do this.