fix: type correctness around symbols and context variables
This diff adds type annotations and updates the context variables
code to make the expectations explicity regarding when we want
symbols, and when we want context variables.

Additionally, we change the molveno example to include a more
complex weather situation, to exercise the previous bug.

The related issue is https://github.com/fbk-most/dt-model/issues/37.

If this diff is accepted, I'll add more tests either to this PR
or to a follow up PR, depending on how we proceed.