refactor: rename {,civic_digital_twins}/dt_model (#42)
This diff creates a toplevel package named civic_digital_twins and moves
the dt_model under the civic_digital_twins package.

This change makes the package name aligned with the pypi package name
and the repository name, for consistency.

While there, be ready for the next release and crank the minor number
wheel.