feat(fronted): add DAG forest partitioning (#72)
This diff introduces forest partitioning on the DAG. We overlay a forest
over the DAG, where the root of each tree is a specific node to compute.
We produce a list of topologically sorted trees where each tree is
rooted into one of the given nodes. The tree distinguishes between
*input* nodes (i.e., nodes that should be known before evaluating the
tree) and *body* nodes (i.e., topologically sorted nodes to evaluate the
tree). This specific view is definitely useful for debugging and
understanding the overall structure of the DAG by isolating the various
components.

As things stand, I am doing this in the context of
https://github.com/fbk-most/civic-digital-twins/issues/58.