feat(graph): allow setting node name if not set (#77)
Empirical experience trying to debug the produced models suggests that
being able to set the name of a node if not already set is a good
primitive for assigning a name to constraints usages. Sure, we're losing
a bit of purity, because we mutate the node, but the DX is better,
therefore it makes sense to introduce this change.

Part of https://github.com/fbk-most/civic-digital-twins/issues/58