fix(dt_model): make debug output easier to understand
This diff includes several small changes meant to make the debug
output easier to understand for a developer. Namely:

1. Represent symbol values as string, which intuitively makes
sense and reduces the cognitive load of reading the logs.

2. Automatically assign a name to SymIndex nodes if it was not
already set, which helps to link nodes to the model.

3. Specifically, assign `f'{name}_usage'` to the constraint
usage. This is one of the most important indexes and it's
really helpful to locate it inside the code dump and the code
execution logs.

With these changes implemented, I think I have done enough
for making the overal debugging story better for now.

Closes https://github.com/fbk-most/civic-digital-twins/issues/58