fix(forest): add constants as deps (#87)
One of the goals of producing a forest is to isolate a subtree and
explicitly name all the leaves it depends on. In terms of potentially
generating code using Numba (as described in the
https://github.com/fbk-most/civic-digital-twins/issues/54 issue), this
design is optimal, because it allows to neatly separate the algorithm
proper and its dependencies. Beyond Numba, with this change, we also get
easier to understand code where we do not instantiate the same constants
over and over again. As such, with this change we also improve the
debuggability of the engine as documented by
https://github.com/fbk-most/civic-digital-twins/issues/58.