feat(numpybackend): evaluate an ir.DAG (#89)
With this change, we've successfully integrated the generic concept of
an `ir.DAG` into the while compiler engine.

This means we're now very well positioned in this or other backends to
act upon the initial assignments to deal with string dtypes. If the
backend directly executes an `ir.DAG`, in fact, it can easily transform
any initial assignment containing strings to numbers by using an
internal translation map. Additionally, the compiler can easily emit the
proper diagnostics explaining its replacements, which helps
debuggability.

Part of https://github.com/fbk-most/civic-digital-twins/issues/84 and
https://github.com/fbk-most/civic-digital-twins/issues/58.