feat(executor): dump round-tripping code
This diff introduces minimal JIT-related code to tweak the executor
so that the debug trace emitted is valid NumPy code allowing to re-run
the same model instance once again.

I suspect this is the maximum debugging and transparency level to
which we can get in this short period of time.

I added some supporting code to `jit.py` since the functionality we're
using for debugging is the same we'd need to implement a JIT.

Part of https://github.com/fbk-most/civic-digital-twins/issues/58.