feat(numpybackend): add infra to dump NumPy code (#93)
This diff adds infrastructure to correctly dump NumPy code generated
from a `graph.Node`. Technically, this code is also the basic foundation
to implement JIT compiling, but I am not going down this road in this
iteration. What we are adding here is the bare minimum to correctly
transform a `graph.Node` to equivalent NumPy-based Python AST along with
extra code for unparsing and dumping the resulting source code.

With this diff committed, most of the work required by
https://github.com/fbk-most/civic-digital-twins/issues/58 has now been
done, and we're really well positioned to have clear and round-trippable
NumPy based debug output.