feat(numpybackend): dump round-tripping code (#92)
This diff ensures the numpybackend emits round-tripping, valid NumPy
code. In other words, the debugging trace produced by a model is itself
a re-runnable, concrete model (save for the need to add `import numpy as
np` at the top of the generated trace, which is a minor issue IMHO).

I suspect this is the maximum debugging and transparency level to which
we can get in this short period of time.

I added some supporting code to `jit.py` since the functionality we're
using for debugging is the same we'd need to implement a JIT.
Specifically, I added the functionality to transform a graph.Node to the
corresponding Python AST expression, which is also the basic building
block required to implement JIT coding with Numba.

Part of https://github.com/fbk-most/civic-digital-twins/issues/58.