doc(engine): add design document (#65)
This design document explains the rationale, design, and implementation
of the engine package, which implements a compiler for the internal DSL
used to express computation graphs generated by the `dt_model` package.

Note that this commit does not mean to include the *final* and *perfect*
design document, rather it contains an initial draft, to be merged in
tree and to be further refined in collaboration with the rest of the
team.

Closes https://github.com/fbk-most/civic-digital-twins/issues/57.